Fixed a small variable assignment bug that was still causing persistent geocoder issues (also turned a select() into a selectAll() just in case, since the bug was causing multiple instances). Did a bunch of refactoring to cut out about a third of the size of the massive Map.svelte file. Tried to move things that were not directly related to the state of the component out, probably a little more that can be done, but it's probably going to be a ~1000 line component regardless.
